Completed the 'better cd' behavior
We're using zoxide to do `cd`. However, if you enter something
that is not in the current directory, the default oxide behavior
is just to jump you to the best match in its database silently.

This is not really okay. The new implemented behavior is as follows:

If the user provides an exact match, zoxide is used to `cd`.

If the user provides a string with a partial match with one of the
directory names, we `cd` into that directory.

If the use rprovides a string with multiple partial matches with
one of the directory names, fzf is used to let the user select from
one of the matched results.

If there are *no* matches, then we just inform the user the directory
does not exist.